Renamed+refactored source_cog
